skin: customise home page panels
