+++ skin/src/components/HomePage/Features.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+const Features = () => (
+  <section>
+    <h4>A few things you can do here</h4>
+    <ul>
+      <li>See recent log events for your apps.</li>
+      <li>Create and bind service instances for your apps.</li>
+      <li>Create and bind routes for your apps.</li>
+      <li>Change memory allocation and number of instances for your apps.</li>
+      <li>Restart your apps.</li>
+      <li>Manage permissions for users of your orgs and spaces.</li>
+    </ul>
+  </section>
+);
+
+export default Features;
